fix: remove duplicate span/pos impls, rename span test feature gate
