+++ src/de/tests.rs
@@ -15,23 +15,29 @@ use crate::{
     value::Number,
 };
 
+#[cfg(feature = "internal-span-substring-test")]
+use crate::util::span_substring::check_error_span_inclusive;
+
+#[cfg(feature = "internal-span-substring-test")]
+use crate::util::span_substring::check_error_span_exclusive;
+
 #[derive(Debug, PartialEq, Deserialize)]
 struct EmptyStruct1;
 
 #[derive(Debug, PartialEq, Deserialize)]
 struct EmptyStruct2 {}
 
-#[derive(Debug, PartialEq, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Deserialize)]
 struct NewType(i32);
 
-#[derive(Debug, PartialEq, Deserialize)]
+#[derive(Clone, Copy, Debug, PartialEq, Deserialize)]
 #[serde(rename = "")]
 struct UnnamedNewType(i32);
 
 #[derive(Debug, PartialEq, Deserialize)]
 struct TupleStruct(f32, f32);
 
-#[derive(Debug, PartialEq, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Deserialize)]
 #[serde(rename = "")]
 struct UnnamedTupleStruct(f32, f32);
 
@@ -72,28 +78,18 @@ fn test_struct() {
     check_from_str_bytes_reader("NewType(42)", Ok(NewType(42)));
     check_from_str_bytes_reader("(33)", Ok(NewType(33)));
 
-    check_from_str_bytes_reader::<NewType>(
-        "NewType",
-        Err(SpannedError {
-            code: Error::ExpectedNamedStructLike("NewType"),
-            span: Span {
-                start: Position { line: 1, col: 1 },
-                end: Position { line: 1, col: 8 },
-            },
-        }),
-    );
+    let bogus_struct = "NewType";
+    let expected_err = Err(SpannedError {
+        code: Error::ExpectedNamedStructLike("NewType"),
+        span: Span {
+            start: Position { line: 1, col: 1 },
+            end: Position { line: 1, col: 8 },
+        },
+    });
+    check_from_str_bytes_reader::<NewType>(bogus_struct, expected_err.clone());
 
-    check_error_span_exclusive::<NewType>(
-        "NewType",
-        Err(SpannedError {
-            code: Error::ExpectedNamedStructLike("NewType"),
-            span: Span {
-                start: Position { line: 1, col: 1 },
-                end: Position { line: 1, col: 8 },
-            },
-        }),
-        "NewType",
-    );
+    #[cfg(feature = "internal-span-substring-test")]
+    check_error_span_exclusive::<NewType>(bogus_struct, expected_err, "NewType");
 
     check_from_str_bytes_reader::<UnnamedNewType>(
         "",
@@ -107,27 +103,19 @@ fn test_struct() {
     );
     check_from_str_bytes_reader("(33)", Ok(UnnamedNewType(33)));
 
-    check_from_str_bytes_reader::<UnnamedNewType>(
-        "Newtype",
-        Err(SpannedError {
-            code: Error::ExpectedNamedStructLike(""),
-            span: Span {
-                start: Position { line: 1, col: 1 },
-                end: Position { line: 1, col: 8 },
-            },
-        }),
-    );
-    check_error_span_exclusive::<UnnamedNewType>(
-        "Newtype",
-        Err(SpannedError {
-            code: Error::ExpectedNamedStructLike(""),
-            span: Span {
-                start: Position { line: 1, col: 1 },
-                end: Position { line: 1, col: 8 },
-            },
-        }),
-        "Newtype",
-    );
+    let bogus_struct = "NewType";
+    let expected_err = Err(SpannedError {
+        code: Error::ExpectedNamedStructLike(""),
+        span: Span {
+            start: Position { line: 1, col: 1 },
+            end: Position { line: 1, col: 8 },
+        },
+    });
+
+    check_from_str_bytes_reader::<UnnamedNewType>(bogus_struct, expected_err.clone());
+
+    #[cfg(feature = "internal-span-substring-test")]
+    check_error_span_exclusive::<UnnamedNewType>(bogus_struct, expected_err, "NewType");
 
     check_from_str_bytes_reader("TupleStruct(2,5,)", Ok(TupleStruct(2.0, 5.0)));
     check_from_str_bytes_reader("(3,4)", Ok(TupleStruct(3.0, 4.0)));
@@ -142,28 +130,18 @@ fn test_struct() {
         }),
     );
 
-    check_from_str_bytes_reader::<UnnamedTupleStruct>(
-        "TupleStruct(2,5,)",
-        Err(SpannedError {
-            code: Error::ExpectedNamedStructLike(""),
-            span: Span {
-                start: Position { line: 1, col: 1 },
-                end: Position { line: 1, col: 12 },
-            },
-        }),
-    );
+    let bogus_struct = "TupleStruct(2,5,)";
+    let expected_err = Err(SpannedError {
+        code: Error::ExpectedNamedStructLike(""),
+        span: Span {
+            start: Position { line: 1, col: 1 },
+            end: Position { line: 1, col: 12 },
+        },
+    });
+    check_from_str_bytes_reader::<UnnamedTupleStruct>(bogus_struct, expected_err.clone());
 
-    check_error_span_exclusive::<UnnamedTupleStruct>(
-        "TupleStruct(2,5,)",
-        Err(SpannedError {
-            code: Error::ExpectedNamedStructLike(""),
-            span: Span {
-                start: Position { line: 1, col: 1 },
-                end: Position { line: 1, col: 12 },
-            },
-        }),
-        "TupleStruct",
-    );
+    #[cfg(feature = "internal-span-substring-test")]
+    check_error_span_exclusive::<UnnamedTupleStruct>(bogus_struct, expected_err, "TupleStruct");
 
     check_from_str_bytes_reader("(3,4)", Ok(UnnamedTupleStruct(3.0, 4.0)));
     check_from_str_bytes_reader::<UnnamedTupleStruct>(
@@ -473,25 +451,21 @@ fn test_err_wrong_value() {
     );
     check_from_str_bytes_reader::<bool>("notabool", err(Error::ExpectedBoolean, (1, 1), (1, 1)));
 
-    check_from_str_bytes_reader::<MyStruct>(
-        "MyStruct(\n    x: true)",
-        err(Error::ExpectedFloat, (2, 7), (2, 8)),
-    );
-    check_error_span_inclusive::<MyStruct>(
-        "MyStruct(\n    x: true)",
-        err(Error::ExpectedFloat, (2, 7), (2, 8)),
-        " t",
-    );
+    let bogus_struct = "MyStruct(\n    x: true)";
+    let expected_err = err(Error::ExpectedFloat, (2, 7), (2, 8));
 
-    check_from_str_bytes_reader::<MyStruct>(
-        "MyStruct(\n    x: 3.5, \n    y:)",
-        err(Error::ExpectedFloat, (3, 7), (3, 7)),
-    );
-    check_error_span_inclusive::<MyStruct>(
-        "MyStruct(\n    x: 3.5, \n    y:)",
-        err(Error::ExpectedFloat, (3, 7), (3, 7)),
-        ")",
-    );
+    check_from_str_bytes_reader::<MyStruct>(bogus_struct, expected_err.clone());
+
+    #[cfg(feature = "internal-span-substring-test")]
+    check_error_span_inclusive::<MyStruct>(bogus_struct, expected_err, " t");
+
+    let bogus_struct = "MyStruct(\n    x: 3.5, \n    y:)";
+    let expected_err = err(Error::ExpectedFloat, (3, 7), (3, 7));
+
+    check_from_str_bytes_reader::<MyStruct>(bogus_struct, expected_err.clone());
+
+    #[cfg(feature = "internal-span-substring-test")]
+    check_error_span_inclusive::<MyStruct>(bogus_struct, expected_err, ")");
 }
 
 #[test]
@@ -504,7 +478,7 @@ fn test_perm_ws() {
 
 #[test]
 fn untagged() {
-    #[derive(Deserialize, Debug, PartialEq)]
+    #[derive(Deserialize, Clone, Debug, PartialEq)]
     #[serde(untagged)]
     enum Untagged {
         U8(u8),
@@ -520,28 +494,19 @@ fn untagged() {
     // - serde content uses deserialize_any, which retriggers the struct type check
     // - struct type check inside a serde content performs a full newtype check
     // - newtype check fails on the unclosed struct
-    check_from_str_bytes_reader::<Untagged>(
-        "Value(()",
-        Err(crate::error::SpannedError {
-            code: crate::Error::Eof,
-            span: Span {
-                start: Position { line: 1, col: 8 },
-                end: crate::error::Position { line: 1, col: 9 },
-            },
-        }),
-    );
+    //
+    let bogus_struct = "Value(()";
+    let expected_err = Err(crate::error::SpannedError {
+        code: crate::Error::Eof,
+        span: Span {
+            start: Position { line: 1, col: 8 },
+            end: crate::error::Position { line: 1, col: 9 },
+        },
+    });
+    check_from_str_bytes_reader::<Untagged>(bogus_struct, expected_err.clone());
 
-    check_error_span_exclusive::<Untagged>(
-        "Value(()",
-        Err(crate::error::SpannedError {
-            code: crate::Error::Eof,
-            span: Span {
-                start: Position { line: 1, col: 8 },
-                end: crate::error::Position { line: 1, col: 9 },
-            },
-        }),
-        ")",
-    );
+    #[cfg(feature = "internal-span-substring-test")]
+    check_error_span_exclusive::<Untagged>(bogus_struct, expected_err, ")");
 }
 
 #[test]
@@ -560,28 +525,19 @@ fn rename() {
 
 #[test]
 fn forgot_apostrophes() {
-    check_from_str_bytes_reader::<(i32, String)>(
-        "(4, \"Hello)",
-        Err(SpannedError {
-            code: Error::ExpectedStringEnd,
-            span: Span {
-                start: Position { line: 1, col: 5 },
-                end: Position { line: 1, col: 6 },
-            },
-        }),
-    );
+    let bogus_struct = "(4, \"Hello)";
+    let expected_err = Err(SpannedError {
+        code: Error::ExpectedStringEnd,
+        span: Span {
+            start: Position { line: 1, col: 5 },
+            end: Position { line: 1, col: 6 },
+        },
+    });
 
-    check_error_span_exclusive::<(i32, String)>(
-        "(4, \"Hello)",
-        Err(SpannedError {
-            code: Error::ExpectedStringEnd,
-            span: Span {
-                start: Position { line: 1, col: 5 },
-                end: Position { line: 1, col: 6 },
-            },
-        }),
-        "\"",
-    );
+    check_from_str_bytes_reader::<(i32, String)>(bogus_struct, expected_err.clone());
+
+    #[cfg(feature = "internal-span-substring-test")]
+    check_error_span_exclusive::<(i32, String)>(bogus_struct, expected_err, "\"");
 }
 
 #[test]
@@ -594,37 +550,26 @@ fn expected_attribute() {
 
 #[test]
 fn expected_attribute_end() {
-    check_from_str_bytes_reader::<String>(
-        "#![enable(unwrap_newtypes) \"Hello\"",
-        err(Error::ExpectedAttributeEnd, (1, 27), (1, 28)),
-    );
-    check_error_span_inclusive::<String>(
-        "#![enable(unwrap_newtypes) \"Hello\"",
-        err(Error::ExpectedAttributeEnd, (1, 27), (1, 28)),
-        " \"",
-    );
+    let bogus_struct = "#![enable(unwrap_newtypes) \"Hello\"";
+    let expected_err = err(Error::ExpectedAttributeEnd, (1, 27), (1, 28));
+    check_from_str_bytes_reader::<String>(bogus_struct, expected_err.clone());
+
+    #[cfg(feature = "internal-span-substring-test")]
+    check_error_span_inclusive::<String>(bogus_struct, expected_err, " \"");
 }
 
 #[test]
 fn invalid_attribute() {
-    check_from_str_bytes_reader::<String>(
-        "#![enable(invalid)] \"Hello\"",
-        err(
-            Error::NoSuchExtension("invalid".to_string()),
-            (1, 11),
-            (1, 18),
-        ),
+    let bogus_struct = "#![enable(invalid)] \"Hello\"";
+    let expected_err = err(
+        Error::NoSuchExtension("invalid".to_string()),
+        (1, 11),
+        (1, 18),
     );
+    check_from_str_bytes_reader::<String>(bogus_struct, expected_err.clone());
 
-    check_error_span_exclusive::<String>(
-        "#![enable(invalid)] \"Hello\"",
-        err(
-            Error::NoSuchExtension("invalid".to_string()),
-            (1, 11),
-            (1, 18),
-        ),
-        "invalid",
-    );
+    #[cfg(feature = "internal-span-substring-test")]
+    check_error_span_exclusive::<String>(bogus_struct, expected_err, "invalid");
 }
 
 #[test]
@@ -803,63 +748,6 @@ fn check_from_str_bytes_reader<T: serde::de::DeserializeOwned + PartialEq + core
         let res_reader = super::from_reader::<&[u8], T>(ron.as_bytes());
         assert_eq!(res_reader, check);
     }
-}
-
-/// Given a string `ron`, a [`SpannedResult`], and a substring, verify that trying to parse `ron` results in an error
-/// equal to the [`SpannedResult`] with a Span that exclusively (as in `[start..end]`) selects that substring.
-/// Note that there are two versions of this helper, inclusive and exclusive. This is because while the parser cursor
-/// arithmetic that computes span positions always produces exclusive spans (as in `[start..end]`),
-/// when doing validation against a target substring, the inclusive check including the final grapheme that triggered
-/// the error is often a more intuitive target to check against.
-/// Meanwhile, if the parser threw an EOF, for example, there is no final grapheme to check, and so
-/// only the exclusive check would produce a meaningful result.
-fn check_error_span_exclusive<T: serde::de::DeserializeOwned + PartialEq + core::fmt::Debug>(
-    ron: &str,
-    check: SpannedResult<T>,
-    substr: &str,
-) {
-    let res_str = super::from_str::<T>(ron);
-    assert_eq!(res_str, check);
-
-    let res_bytes = super::from_bytes::<T>(ron.as_bytes());
-    assert_eq!(res_bytes, check);
-
-    #[cfg(feature = "std")]
-    {
-        let res_reader = super::from_reader::<&[u8], T>(ron.as_bytes());
-        assert_eq!(res_reader, check);
-    }
-
-    assert_eq!(
-        check.unwrap_err().span.substring_exclusive(ron).unwrap(),
-        substr
-    );
-}
-
-/// Given a string `ron`, a [`SpannedResult`], and a substring, verify that trying to parse `ron` results in an error
-/// equal to the [`SpannedResult`] with a Span that inclusively (as in `[start..=end`]) selects that substring.
-/// See [`check_error_span_exclusive`] for the rationale behind both versions of this helper.
-fn check_error_span_inclusive<T: serde::de::DeserializeOwned + PartialEq + core::fmt::Debug>(
-    ron: &str,
-    check: SpannedResult<T>,
-    substr: &str,
-) {
-    let res_str = super::from_str::<T>(ron);
-    assert_eq!(res_str, check);
-
-    let res_bytes = super::from_bytes::<T>(ron.as_bytes());
-    assert_eq!(res_bytes, check);
-
-    #[cfg(feature = "std")]
-    {
-        let res_reader = super::from_reader::<&[u8], T>(ron.as_bytes());
-        assert_eq!(res_reader, check);
-    }
-
-    assert_eq!(
-        check.unwrap_err().span.substring_inclusive(ron).unwrap(),
-        substr
-    );
 }
 
 #[test]
