+++ src/util/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "internal-span-substring-test")]
+pub mod span_substring;
